added $no_batch global configuration variable for colored exportation of embedded source code
When $no_batch is set to a truthy value, then Emacs will not be run in
batch mode, but will open a graphical client to enable color
exportation through htmlize.el.

Note: as a followup to this change, it might be nice to allow the
optional loading of color themes.